[FIXED] Screen and component names
