+++ package.json
@@ -9,6 +9,7 @@
     "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-icons": "^3.11.0",
     "react-modal": "^3.11.2",
     "react-responsive-modal": "^5.1.0",
     "react-router-dom": "^5.2.0",
