--- src/components/ScreenHome.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const ScreenHome = () => {
-  return <div></div>;
-};
