--- src/components/ScreenNavBar.js
@@ -1,23 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-
-export const NavBar = () => {
-  return (
-    <nav>
-      <ul>
-        <li>
-          <Link to="/">Home</Link>
-        </li>
-        <li>
-          <Link to="/login">Login</Link>
-        </li>
-        <li>
-          <Link to="/register">Register</Link>
-        </li>
-        <li>
-          <Link to="/favorites">Favorites</Link>
-        </li>
-      </ul>
-    </nav>
-  );
-};
