+++ src/routers/AppRouter.js
@@ -1,10 +1,10 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
-import { NavBar } from "../components/ScreenNavBar";
-import { ScreenLoginModal } from "../components/ScreenLoginModal";
+import { NavBar } from "../components/NavBar";
+import { LoginModal } from "../components/LoginModal";
 import { ScreenHome } from "../components/ScreenHome";
-import { ScreenRegisterModal } from "../components/ScreenRegisterModal";
+import { RegisterModal } from "../components/RegisterModal";
 import { Favorites } from "../components/ScreenFavorites";
 import { RoomDetails } from "../components/ScreenRoomDetails";
 
@@ -15,8 +15,8 @@ export const AppRouter = () => {
         <NavBar />
         <Switch>
           <Route exact path="/" component={ScreenHome} />
-          <Route exact path="/login" component={ScreenLoginModal} />
-          <Route exact path="/register" component={ScreenRegisterModal} />
+          <Route exact path="/login" component={LoginModal} />
+          <Route exact path="/register" component={RegisterModal} />
           <Route exact path="/favorites" component={Favorites} />
           <Route exact path="/room-details" component={RoomDetails} />
         </Switch>
